Sorting Algo logic Revised, Input Output Logic
Ok I read up on the specifications more. I was wrong and I think there is no need to follow the format I initially thought. Continue the usual programming of the sorting algorithms.

Some key things to note for the Input Output logic I formulated in Main and DataSetRandomizer

Main accepts a filepath and outputs it by default to the folder "SortResult"

DataSetRandomizer (P.S. if you use VS code, make sure to click File > Open Folder > <Folder of DataSetRandomizer which is "Make New Data">)

DataSetRandomizer just asks for two things, the filename of your file and the number of data entries it should contain. It would then randomize and combine names and select random numbers and give out the result in the folder named "OutputData"